+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added statistics compatibility trait `StatRec` for generalizing across different
   versions of the statistics record
 - Added `AsRef<[u8]>` implementations for `RecordEnum` and `RecordRefEnum`
+- Added new off-market publishers for Eurex, and European Energy Exchange (EEX)
 
 ### Breaking changes
 - Made `Record` a subtrait of `AsRef<[u8]>` as all records should be convertible to